requirements.txt
google-genai
scikit-learn
python-dotenv